SUBMARINE-890. Extract all resource creation spec
