+++ submarine-cloud-v2/main.go
@@ -21,6 +21,7 @@ import (
 	"flag"
 	clientset "github.com/apache/submarine/submarine-cloud-v2/pkg/client/clientset/versioned"
 	informers "github.com/apache/submarine/submarine-cloud-v2/pkg/client/informers/externalversions"
+	"github.com/apache/submarine/submarine-cloud-v2/pkg/controller"
 	"github.com/apache/submarine/submarine-cloud-v2/pkg/signals"
 	"os"
 	"time"
@@ -84,7 +85,7 @@ func main() {
 	//       ex: namespace informer
 
 	// Create a Submarine operator
-	controller := NewController(incluster, kubeClient, submarineClient, traefikClient,
+	submarineController := controller.NewController(incluster, kubeClient, submarineClient, traefikClient,
 		kubeInformerFactory.Core().V1().Namespaces(),
 		kubeInformerFactory.Apps().V1().Deployments(),
 		kubeInformerFactory.Core().V1().Services(),
@@ -104,7 +105,7 @@ func main() {
 	traefikInformerFactory.Start(stopCh)
 
 	// Run controller
-	if err = controller.Run(1, stopCh); err != nil {
+	if err = submarineController.Run(1, stopCh); err != nil {
 		klog.Fatalf("Error running controller: %s", err.Error())
 	}
 }
